I did a small benchmark to test the change:
```julia
include("src/AutoGrad.jl")
using BenchmarkTools
f(x) = sum(x->x^2, x)
v = rand(100)

BenchmarkTools.DEFAULT_PARAMETERS.seconds = 20.
BenchmarkTools.DEFAULT_PARAMETERS.time_tolerance = 0.01
BenchmarkTools.DEFAULT_PARAMETERS.samples = 50000
@benchmark AutoGrad.grad(f)(v)
```
performance are essentially the same

before
```julia
BenchmarkTools.Trial: 
  memory estimate:  383.31 KiB
  allocs estimate:  9198
  --------------
  minimum time:     1.243 ms (0.00% GC)
  median time:      1.310 ms (0.00% GC)
  mean time:        1.435 ms (7.18% GC)
  maximum time:     10.164 ms (84.91% GC)
  --------------
  samples:          13852
  evals/sample:     1
  time tolerance:   1.00%
  memory tolerance: 1.00%
```
after 
```julia
BenchmarkTools.Trial: 
  memory estimate:  383.31 KiB
  allocs estimate:  9198
  --------------
  minimum time:     1.146 ms (0.00% GC)
  median time:      1.247 ms (0.00% GC)
  mean time:        1.352 ms (5.78% GC)
  maximum time:     8.805 ms (71.69% GC)
  --------------
  samples:          14686
  evals/sample:     1
  time tolerance:   1.00%
  memory tolerance: 1.00%
```