@denizyuret 2017-04-02
I was able to replicate your results with Julia5.  In Julia4 map is 4x slower, however.  I am pushing a conditional fix in the latest master.