@CarloLucibello 2017-04-02
All right. In any case when you will make the upgrade to julia 0.6, soon to be released, support for 0.4 will have to be dropped (keeping compatibility with three versions is an unnecessary nightmare)